+++ packages/compat/config/chain.ts
@@ -1,8 +1,18 @@
+import { ParachainId } from '@zenlink-interface/chain'
+
 export {
-  AMM_ENABLED_NETWORKS,
   EVM_NETWORKS,
   isEvmNetwork,
   isSubstrateNetwork,
   SUBSTRATE_NETWORKS,
-  SUPPORTED_CHAIN_IDS,
 } from '@zenlink-interface/chain'
+
+export const AMM_ENABLED_NETWORKS = [
+  ParachainId.ASTAR,
+  ParachainId.MOONRIVER,
+  ParachainId.MOONBEAM,
+  ParachainId.BIFROST_KUSAMA,
+  ParachainId.BIFROST_POLKADOT,
+]
+
+export const SUPPORTED_CHAIN_IDS = Array.from(new Set([...AMM_ENABLED_NETWORKS]))
